+++ psec/__init__.py
@@ -4,13 +4,13 @@ import pbr.version
 
 version_info = pbr.version.VersionInfo('psec')
 try:
-    __version__ = version_info.version_string()
-except AttributeError:
+    __version__ = version_info.cached_version_string()
+except Exception:
     __version__ = '19.4.4'
 
 try:
     __release__ = version_info.release_string()
-except AttributeError:
+except Exception:
     __release__ = '19.4.4'
 
 __author__ = 'Dave Dittrich'